elearning_platform/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
pika==1.3.2
python-consul==1.1.0
requests==2.31.0
psycopg2-binary==2.9.9
